Added TestMatrix and AbstractPositioner
These struct are used to define and control actuators on an experiment.

TestMatrix provides the points (configurations of the experiments) and
AbstractPositioner is a generic interface that handles the movement.

TestMatrix objects can be combined to form cartesian products of different
points. The same happens with the product of AbstractPositioner objects.